Made SelectModerationStatus as a reusable component
